requirements.txt
# ── Core ───────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
sse-starlette>=2.2.0

# ── Agent ──────────────────────────────────────────────────────
pydantic-ai>=0.2.0
google-genai>=1.0.0
anthropic>=0.40.0

# ── Database ───────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
pgvector>=0.3.6
alembic>=1.14.0

# ── Cache ──────────────────────────────────────────────────────
redis>=5.2.0

# ── RAG / Crawler ─────────────────────────────────────────────
httpx[http2]>=0.28.0
beautifulsoup4>=4.12.0
pypdf>=5.0.0

# ── Auth ───────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ── Utils ──────────────────────────────────────────────────────
python-dotenv>=1.0.0
python-multipart>=0.0.18
jinja2>=3.1.0
